modified data types

+++ src/BacklightColors.cpp
@@ -8,22 +8,22 @@
 
 #include "BacklightColors.h"
 
-const int BacklightColors::BLACK[3] = {0, 0, 0};         ///< Black color (0, 0, 0).
-const int BacklightColors::WHITE[3] = {255, 255, 255};   ///< White color (255, 255, 255).
-const int BacklightColors::HOTPINK[3] = {255, 10, 127};  ///< Hot Pink color (255, 10, 127).
-const int BacklightColors::RED[3] = {255, 10, 10};       ///< Red color (255, 10, 10).
-const int BacklightColors::ORANGE[3] = {255, 127, 10};   ///< Orange color (255, 127, 10).
-const int BacklightColors::YELLOW[3] = {255, 255, 10};   ///< Yellow color (255, 255, 10).
-const int BacklightColors::LIME[3] = {127, 255, 10};     ///< Lime Green color (127, 255, 10).
-const int BacklightColors::GREEN[3] = {10, 255, 10};     ///< Green color (10, 255, 10).
-const int BacklightColors::SPRING[3] = {10, 255, 127};   ///< Spring Green color (10, 255, 127).
-const int BacklightColors::CYAN[3] = {10, 255, 255};     ///< Cyan color (10, 255, 255).
-const int BacklightColors::SKY[3] = {10, 127, 255};      ///< Sky Blue color (10, 127, 255).
-const int BacklightColors::BLUE[3] = {10, 10, 255};      ///< Blue color (10, 10, 255).
-const int BacklightColors::VIOLET[3] = {127, 10, 255};   ///< Violet color (127, 10, 255).
-const int BacklightColors::MAGENTA[3] = {255, 10, 255};  ///< Magenta color (255, 10, 255).
-const int BacklightColors::GRAY[3] = {128, 128, 128};    ///< Gray color (128, 128, 128).
-const int BacklightColors::GREY[3] = {128, 128, 128};    ///< Grey color (128, 128, 128).
-const int BacklightColors::OLIVE[3] = {128, 128, 10};    ///< Olive color (128, 128, 10).
-const int BacklightColors::TEAL[3] = {10, 128, 128};     ///< Teal color (10, 128, 128).
-const int BacklightColors::PURPLE[3] = {128, 10, 128};   ///< Purple color (128, 10, 128).
+const uint8_t BacklightColors::BLACK[3] = {0, 0, 0};         ///< Black color (0, 0, 0).
+const uint8_t BacklightColors::WHITE[3] = {255, 255, 255};   ///< White color (255, 255, 255).
+const uint8_t BacklightColors::HOTPINK[3] = {255, 10, 127};  ///< Hot Pink color (255, 10, 127).
+const uint8_t BacklightColors::RED[3] = {255, 10, 10};       ///< Red color (255, 10, 10).
+const uint8_t BacklightColors::ORANGE[3] = {255, 127, 10};   ///< Orange color (255, 127, 10).
+const uint8_t BacklightColors::YELLOW[3] = {255, 255, 10};   ///< Yellow color (255, 255, 10).
+const uint8_t BacklightColors::LIME[3] = {127, 255, 10};     ///< Lime Green color (127, 255, 10).
+const uint8_t BacklightColors::GREEN[3] = {10, 255, 10};     ///< Green color (10, 255, 10).
+const uint8_t BacklightColors::SPRING[3] = {10, 255, 127};   ///< Spring Green color (10, 255, 127).
+const uint8_t BacklightColors::CYAN[3] = {10, 255, 255};     ///< Cyan color (10, 255, 255).
+const uint8_t BacklightColors::SKY[3] = {10, 127, 255};      ///< Sky Blue color (10, 127, 255).
+const uint8_t BacklightColors::BLUE[3] = {10, 10, 255};      ///< Blue color (10, 10, 255).
+const uint8_t BacklightColors::VIOLET[3] = {127, 10, 255};   ///< Violet color (127, 10, 255).
+const uint8_t BacklightColors::MAGENTA[3] = {255, 10, 255};  ///< Magenta color (255, 10, 255).
+const uint8_t BacklightColors::GRAY[3] = {128, 128, 128};    ///< Gray color (128, 128, 128).
+const uint8_t BacklightColors::GREY[3] = {128, 128, 128};    ///< Grey color (128, 128, 128).
+const uint8_t BacklightColors::OLIVE[3] = {128, 128, 10};    ///< Olive color (128, 128, 10).
+const uint8_t BacklightColors::TEAL[3] = {10, 128, 128};     ///< Teal color (10, 128, 128).
+const uint8_t BacklightColors::PURPLE[3] = {128, 10, 128};   ///< Purple color (128, 10, 128).
